+++ module/src/main/java/com/inseye/shared/utils/BindingDiedDelegate.java
@@ -11,6 +11,11 @@ package com.inseye.shared.utils;
 
 import android.content.ComponentName;
 
+/**
+ * @deprecated
+ * This interface is not mandatory part of the library and will be removed.
+ */
+@Deprecated(forRemoval = true)
 public interface BindingDiedDelegate {
     void onBindingDied(ComponentName name);
 }