+++ module/src/main/java/com/inseye/shared/utils/ServiceDisconnectedDelegate.java
@@ -10,7 +10,11 @@
 package com.inseye.shared.utils;
 
 import android.content.ComponentName;
-
+/**
+ * @deprecated
+ * This interface is not mandatory part of the library and will be removed.
+ */
+@Deprecated(forRemoval = true)
 public interface ServiceDisconnectedDelegate {
     void onServiceDisconnected(ComponentName name);
 }